+++ src/components/ui/capsule/constants/capsule.ts
@@ -1,0 +1,26 @@
+export const capsuleColors = [
+  "#FF8BA0",
+  "#6B8AFF",
+  "#B8F0D9",
+  "#FFF4C4",
+  "#FFD4B8",
+  "#D4C4FF",
+];
+
+export const messages = [
+  "오늘은 이곳이 당신을 기다려요!",
+  "새로운 경험이 당신을 반겨요!",
+  "특별한 순간을 만날 시간이에요!",
+  "행운의 팝업을 발견했어요!",
+  "이번 주 놓치면 안 되는 곳!",
+  "당신의 취향 저격 팝업!",
+];
+
+export const categories = [
+  "Character",
+  "Goods",
+  "Exhibition",
+  "Beauty",
+  "Food",
+  "Fashion",
+] as const;
